requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
PyMuPDF==1.24.9
openai==1.26.0
python-dotenv==1.0.1
python-multipart==0.0.5
azure-cosmos==4.7.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1